rename directory util -> utilities for consistency
