+++ include/exadg/incompressible_navier_stokes/spatial_discretization/operator_coupled.cpp
@@ -26,7 +26,7 @@
 #include <exadg/solvers_and_preconditioners/preconditioners/block_jacobi_preconditioner.h>
 #include <exadg/solvers_and_preconditioners/preconditioners/inverse_mass_preconditioner.h>
 #include <exadg/solvers_and_preconditioners/preconditioners/jacobi_preconditioner.h>
-#include <exadg/solvers_and_preconditioners/util/check_multigrid.h>
+#include <exadg/solvers_and_preconditioners/utilities/check_multigrid.h>
 
 namespace ExaDG
 {
