+++ include/exadg/operators/operator_base.h
@@ -41,7 +41,7 @@
 #include <exadg/solvers_and_preconditioners/preconditioners/enum_types.h>
 #include <exadg/solvers_and_preconditioners/solvers/enum_types.h>
 #include <exadg/solvers_and_preconditioners/solvers/wrapper_elementwise_solvers.h>
-#include <exadg/solvers_and_preconditioners/util/invert_diagonal.h>
+#include <exadg/solvers_and_preconditioners/utilities/invert_diagonal.h>
 
 #include <exadg/operators/elementwise_operator.h>
 #include <exadg/operators/integrator_flags.h>
